Rerender the code
