+++ src/index.html
@@ -11,7 +11,7 @@
     <script nomodule src="https://unpkg.com/ionicons@5.5.2/dist/ionicons/ionicons.js"></script>
     <script src="https://kit.fontawesome.com/f654a6a150.js" crossorigin="anonymous"></script>
   </head>
-  <body>
+  <body class="body">
     <div id="list-container" class="list-container">
       <div class="title">
         Today's To Do
