+++ src/removeCompletedTasks.js
@@ -1,14 +1,14 @@
-/* eslint-disable import/extensions */
-import updateStorage from './updateStorage.js';
-import showList from './showList.js';
-import orderTasks from './orderTasks.js';
+// /* eslint-disable import/extensions */
+// import updateStorage from './updateStorage.js';
+// import showList from './showList.js';
+// import orderTasks from './orderTasks.js';
 
-const removeCompletedTasks = (tasks) => {
-  tasks = tasks.filter((task) => task.completed === false);
-  orderTasks(tasks);
-  updateStorage(tasks);
-  showList(tasks);
-  return tasks;
-};
+// const removeCompletedTasks = (tasks) => {
+//   tasks = tasks.filter((task) => task.completed === false);
+//   orderTasks(tasks);
+//   updateStorage(tasks);
+//   showList(tasks);
+//   return tasks;
+// };
 
-export { removeCompletedTasks as default };
+// export { removeCompletedTasks as default };
