+++ src/updateStorage.js
@@ -1,5 +1,5 @@
-const updateStorage = (storage) => {
-  localStorage.setItem('storage', JSON.stringify(storage));
-};
+// const updateStorage = (storage) => {
+//   localStorage.setItem('storage', JSON.stringify(storage));
+// };
 
-export { updateStorage as default };
+// export { updateStorage as default };
